integrate into plandb_methods
